% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{create_columns}
\alias{create_columns}
\title{Create column definitions from data}
\usage{
create_columns(data, editor = FALSE, filter = FALSE)
}
\arguments{
\item{data}{(data.frame) data}

\item{editor}{(bool): Whether to make columns editable.}

\item{filter}{(bool): Whether to add a header filter to the columns.}
}
\description{
Create column definitions from data
}
