% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_color}
\alias{formatter_color}
\title{Set color formatter}
\usage{
formatter_color()
}
\value{
A list of column options for setting up the formatter.
}
\description{
Set color formatter
}
\examples{
data <- data.frame(
  Label = c("R", "G", "B"),
  Color = c("red", "green", "blue")
)

tabulator(data, width = 200) |>
  set_columns(columns = "Color", formatter = formatter_color())
}
