% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_datetime}
\alias{formatter_datetime}
\title{Set datetime formatter}
\usage{
formatter_datetime(
  input_format = "yyyy-MM-dd hh:ss:mm",
  output_format = "yy/MM/dd",
  invalid_placeholder = "(invalid datetime)",
  timezone = NA
)
}
\arguments{
\item{input_format}{(character): The datetime input format.}

\item{output_format}{(character): The datetime output format.}

\item{invalid_placeholder}{(character): The value to be displayed
if an invalid datetime is provided.}

\item{timezone}{(character): The timezone of the datetime.}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Set datetime formatter
}
\details{
To use this formatter, you need to include
the \href{https://moment.github.io/luxon/}{luxon} HTML dependency with \code{tabulator(..., luxon = TRUE)}.
}
\examples{
data <- data.frame(
  Person = c("Franz", "Ferdinand"),
  Birthday = c(
    "2024-12-06 22:00:10",
    "2023-06-07 14:12:45"
  )
)

tabulator(data, luxon = TRUE) |>
  set_columns("Birthday", formatter = formatter_datetime(output_format = "dd.MM.yyyy"))
}
