% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_html}
\alias{formatter_html}
\title{HTML formatter}
\usage{
formatter_html()
}
\value{
A list of column options for setting up the formatter.
}
\description{
HTML formatter
}
\examples{
data <- data.frame(
  id = c(1, 2, 3, 4, 5, 6, 7),
  text_style = c(
    "<i>Italic</i>",
    "<b>Bold</b>",
    "<span style='color: green'>Green</span>",
    "<del>Deleted</del>",
    "<small>Small</small>",
    "H<sub>2</sub>O",
    "x<sup>2</sup>"
  )
)

tabulator(data, width = 400) |>
  set_columns("text_style", formatter = formatter_html())
}
