% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_image}
\alias{formatter_image}
\title{Image formatter}
\usage{
formatter_image(
  height = "50px",
  width = "50px",
  url_prefix = NULL,
  url_suffix = NULL
)
}
\arguments{
\item{height}{(character): A CSS value for the height of the image.}

\item{width}{(character): A CSS value for the width of the image.}

\item{url_prefix}{(character): String to add to the start of the cell value
when generating the image src url.}

\item{url_suffix}{(character): String to add to the end of the cell value
when generating the image src url.}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Image formatter
}
\examples{
image_url <- "https://picsum.photos/id/"
image_size <- 100

image_data <- data.frame(
  image = c(88, 98, 102, 201),
  label = c("Pic No. 88", "Pic No. 98", "Pic No. 102", "Pic No 201")
)

tabulator(image_data, tabulator_options(height = "400px")) |>
  set_columns(columns = "image",
              formatter = formatter_image(
                height = image_size,
                width = image_size,
                url_prefix = image_url,
                url_suffix = glue::glue("/{image_size}")
              )
  )
}
