% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_link}
\alias{formatter_link}
\title{Link formatter}
\usage{
formatter_link(
  label_field = NULL,
  url_prefix = NULL,
  url = NULL,
  target = "_blank"
)
}
\arguments{
\item{label_field}{(character): Column to be used as label for the link.}

\item{url_prefix}{(character): Prefix to add to the URL value.}

\item{url}{(JavaScript function): A JavaScript function that return the URL value.
The cell is passed to the function as its first argument.
Use \link[htmlwidgets]{JS} to pass JS code.}

\item{target}{(character): Target attribute of the anchor tag.}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Link formatter
}
\examples{
data <- data.frame(
  label = c("This is a link", "This is another link"),
  link = c("https://eoda.de", "https://eoda.de/data-science"),
  link2 = c("eoda.de", "github.com"),
  link3 = c("posit-partner", "data-science")
)

js_func <- "(cell) => `https://eoda.de/${cell.getValue()}`"

tabulator(data) |>
  set_columns("link" , formatter= formatter_link(label_field = "label")) |>
  set_columns("link2", formatter= formatter_link(url_prefix = "https://"))|>
  set_columns("link3", formatter= formatter_link(url = htmlwidgets::JS(js_func),
                                                 label_field = "label"))
}
