% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_money}
\alias{formatter_money}
\title{Money formatter}
\usage{
formatter_money(
  decimal = c(",", "."),
  thousand = c(".", ","),
  symbol = "$ ",
  symbol_after = "p",
  negative_sign = "-",
  precision = FALSE
)
}
\arguments{
\item{decimal}{(character): Symbol to represent the decimal point.}

\item{thousand}{(character, bool): Symbol to represent the thousands separator.
Set to \code{FALSE} to disable the separator.}

\item{symbol}{(character): The currency symbol.}

\item{symbol_after}{(bool): Whether to put the symbol after the number.}

\item{negative_sign}{(character, bool): The sign to show in front of the number.
Set to \code{TRUE} causes negative numbers to be enclosed in brackets (123.45),
which is the standard style for negative numbers in accounting.}

\item{precision}{(integer, bool): The number of decimals to display.
Set to \code{FALSE} to display all decimals that are provided.}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Money formatter
}
\examples{
data <- data.frame(
  Account_Number = c(
    123456,
    345667,
    234567,
    234566
  ),
  Account_Balance = c(100, -50, 200.30, -21.5)
)

tabulator(data) |>
  set_columns(
    "Account_Balance",
    formatter = formatter_money(
      symbol = " \U20AC",
      symbol_after = TRUE
    )
  )
}
