% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_star}
\alias{formatter_star}
\title{Star rating formatter}
\usage{
formatter_star(number_of_stars = NA)
}
\arguments{
\item{number_of_stars}{The maximum number of stars to be displayed.
If set to \code{NA}, the maximum value of the column is used.}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Star rating formatter
}
\examples{
data <- data.frame(
  Passengers = c("Hans", "Franz", "Ferdinand", "Julia"),
  PassengerClass = c(1, 2, 1, 3)
)

tabulator(data, width = 200) |>
  set_columns("PassengerClass",
              formatter = formatter_star(number_of_stars = max(data$PassengerClass)))

}
