% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_textarea}
\alias{formatter_textarea}
\title{Text area formatter}
\usage{
formatter_textarea()
}
\value{
A list of column options for setting up the formatter.
}
\description{
Text area formatter
}
\examples{
data <- data.frame(
  id = c(1, 2),
  txt = c(
    "This\nis\nsome\ntext\nwith\nmultiple\nline\nbreaks",
    "- R\n- Python\n- Julia"
  )
)

tabulator(data, width = 200) |>
  set_columns("txt", formatter = formatter_textarea())
}
