% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_tick_cross}
\alias{formatter_tick_cross}
\title{Tick cross formatter}
\usage{
formatter_tick_cross()
}
\value{
A list of column options for setting up the formatter.
}
\description{
Tick cross formatter
}
\examples{
data <- data.frame(
  Artist = c("Art Blackey", "Nirvana", "Bob Marley"),
  Grunge = c(0, 1, 0),
  Jazz = c(1, 0, 0),
  Reggae = c(0, 0, 1)
)

tabulator(data, width = 400) |>
  set_columns("Grunge", formatter = formatter_tick_cross()) |>
  set_columns("Jazz", formatter = formatter_tick_cross()) |>
  set_columns("Reggae", formatter = formatter_tick_cross())
}
