% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_traffic_light}
\alias{formatter_traffic_light}
\title{Set traffic light formatter}
\usage{
formatter_traffic_light(
  min = NA,
  max = NA,
  color = c("green", "orange", "red")
)
}
\arguments{
\item{min}{(numeric): The minimum value for progress bar.
If set to \code{NA}, the minimum value of the column is used.}

\item{max}{(numeric): The maximum value for progress bar.
If set to \code{NA}, the maximum value of the column is used.}

\item{color}{(character): Either a single color or a vector of colors}
}
\value{
A list of column options for setting up the formatter.
}
\description{
Set traffic light formatter
}
\examples{
data <- data.frame(
  label = 1:10,
  value = 1:10
)

tabulator(data, width = 200) |>
  set_columns("value", formatter = formatter_traffic_light())
}
