% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_data_frame}
\alias{list_to_data_frame}
\title{Parse List to Data Frame}
\usage{
list_to_data_frame(x)
}
\arguments{
\item{x}{(list): A list of named lists.}
}
\description{
Parse List to Data Frame
}
