% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{modify_col_def}
\alias{modify_col_def}
\title{Modify column definition}
\usage{
modify_col_def(widget, columns, col_update)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The names of the columns the formatter is applied to.}

\item{col_update}{A named list containing the updates to apply to each column in \code{columns}.
The updates are merged into the existing column definitions.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Modify column definition
}
\examples{

df <- data.frame(values = c(1, 2, 3), names = c("a", "b", "c"))
tabulator(df) |>
  modify_col_def(c("values", "names"),
    col_update = list(hozAlign = "center")
  )
}
