% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{redo}
\alias{redo}
\title{Redo changes}
\usage{
redo(ctx)
}
\arguments{
\item{ctx}{A \code{\link[=tabulatorContext]{tabulatorContext()}} object.}
}
\value{
A \code{\link[=tabulatorContext]{tabulatorContext()}} object
}
\description{
Redo changes
}
\examples{
\donttest{
tabulatorContext("table") |>
  redo()
}
}
