% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_calculation}
\alias{set_calculation}
\title{Set calculation}
\usage{
set_calculation(
  widget,
  column,
  func = c("avg", "max", "min", "sum", "count", "unique"),
  precision = 2,
  pos = c("top", "bottom")
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{column}{(character): The column the \code{func} is applied to.}

\item{func}{(character): The calculation function to be applied
to the values of the \code{column}.}

\item{precision}{(integer)  The number of decimals to display.
Set to \code{FALSE} to display all decimals.}

\item{pos}{(character): Position at which calculated values are displayed.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Set calculation
}
\examples{
tabulator(iris) |>
  set_calculation("Sepal_Length", "avg")
}
