% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_column_editor}
\alias{set_column_editor}
\title{Make columns editable}
\usage{
set_column_editor(widget, columns, type = c("input", "number"))
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{columns}{(character vector): Columns the editor is applied to.}

\item{type}{(character): Either \code{input} or \code{number}.}
}
\description{
Make columns editable
}
\examples{
data <- data.frame(
  age = c(10, 20, 40),
  first_name = c("Franz", "Ferdinand", "Julia"),
  last_name = c("Young", "Bowie", "Blackey")
)

tabulator(data, width = 500) |>
  set_column_editor(c("first_name", "last_name")) |>
  set_column_editor("age", type = "number")
}
