% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{set_column_editor_input}
\alias{set_column_editor_input}
\title{The \strong{input} editor allows entering of a single line of plain text}
\usage{
set_column_editor_input(
  widget,
  columns,
  search = TRUE,
  mask = "",
  select_contents = TRUE,
  elementAttributes = list(maxlength = 10)
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The name of the column the formatter is applied to.}

\item{search}{Use search type input element with clear button}

\item{mask}{Apply a mask to the input to allow characters to be entered only in a certain order}

\item{select_contents}{When the editor is loaded select its text content}

\item{elementAttributes}{Set attributes directly on the input element}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
The \strong{input} editor allows entering of a single line of plain text
}
\examples{
df <- data.frame(values = c(1,2,3),
                 names = c("a","b","c"))
tabulator(df) |>
  set_column_editor_input("names")
}
