% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-editors.R
\name{set_column_editor_numeric}
\alias{set_column_editor_numeric}
\title{The \strong{numeric}  editor allows for numeric entry with a number type input element with increment and decrement buttons.}
\usage{
set_column_editor_numeric(
  widget,
  columns,
  min = 0,
  max = 100,
  step = 1,
  elementAttributes = list(maxlength = 10),
  mask = "",
  select_contents = TRUE,
  vertical_navigation = "table"
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The name of the column the formatter is applied to.}

\item{min}{the maximum allowed value}

\item{max}{the minimum allowed value}

\item{step}{the step size when incrementing/decrementing the value (default 1)}

\item{elementAttributes}{Set attributes directly on the element}

\item{mask}{Apply a mask to the input to allow characters to be entered only in a certain order}

\item{select_contents}{When the editor is loaded select its text content}

\item{vertical_navigation}{determine how use of the up/down arrow keys will affect the editor,
this can take two different types of value: : \emph{editor} and \emph{table}}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
The \strong{numeric}  editor allows for numeric entry with a number type input element with increment and decrement buttons.
}
