% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{set_column_editor_text_area}
\alias{set_column_editor_text_area}
\title{The \strong{textarea} editor allows entering of multiple lines of plain text}
\usage{
set_column_editor_text_area(
  widget,
  columns,
  elementAttributes = list(maxlength = 10),
  mask = "",
  select_contents = TRUE,
  vertical_navigation = "editor",
  shift_enter_submit = TRUE
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The name of the column the formatter is applied to.}

\item{elementAttributes}{Set attributes directly on the textarea element}

\item{mask}{Apply a mask to the input to allow characters to be entered only in a certain order}

\item{select_contents}{When the editor is loaded select its text content}

\item{vertical_navigation}{Determine how use of the up/down arrow keys will affect the editor,
this can take three different types of value: \emph{hybrid}, \emph{editor} and \emph{table}}

\item{shift_enter_submit}{Submit the cell value when the shift and enter keys are pressed}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
The \strong{textarea} editor allows entering of multiple lines of plain text
}
