% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_columns}
\alias{set_columns}
\title{Set column definitions}
\usage{
set_columns(
  widget,
  columns,
  title = NULL,
  formatter = NULL,
  hoz_align = NULL,
  ...
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The names of the columns the formatter is applied to.}

\item{title}{A title that will be displayed for the columns.}

\item{formatter}{A formatter definition as generated by a \verb{formatter_*()} function}

\item{hoz_align}{(character): The horizontal alignment of the column.}

\item{...}{You can add arguments that will be passed on as updates to the column definition.
Checkout the documentation at \url{https://tabulator.info/docs/}}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Set column definitions
}
\examples{
data <- data.frame(
  id = c(1, 2, 3, 4, 5, 6, 7),
  text_style = c(
    "<i>Italic</i>",
    "<b>Bold</b>",
    "<span style='color: green'>Green</span>",
    "<del>Deleted</del>",
    "<small>Small</small>",
    "H<sub>2</sub>O",
    "x<sup>2</sup>"
  )
)

tabulator(data, width = 400) |>
  set_columns("text_style", formatter = formatter_html())
}
