% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_formatter_money}
\alias{set_formatter_money}
\title{Money Formatter}
\usage{
set_formatter_money(
  widget,
  column,
  decimal = ",",
  thousand = ".",
  symbol = "£",
  symbol_after = "p",
  negative_sign = T,
  precision = F,
  hoz_align = "left"
)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the formatter is applied to.}

\item{hoz_align}{(character): The horizontal alignment of the column.}
}
\description{
Money Formatter
}
