% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_formatter_plaintext}
\alias{set_formatter_plaintext}
\title{Set plain text formatter}
\usage{
set_formatter_plaintext(widget, column, hoz_align = "left")
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{column}{The name of the column the formatter is applied to.}

\item{hoz_align}{(character): The horizontal alignment of the column.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Set plain text formatter
}
\examples{
tabulator(iris) |>
  set_formatter_plaintext("Species", hoz_align = "right")
}
