% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_formatter_traffic_light}
\alias{set_formatter_traffic_light}
\title{Traffic Light Formatter}
\usage{
set_formatter_traffic_light(
  widget,
  column,
  min = NA,
  max = NA,
  color = c("green", "orange", "red"),
  hoz_align = "center"
)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the formatter is applied to.}

\item{hoz_align}{(character): The horizontal alignment of the column.}
}
\description{
Traffic Light Formatter
}
