% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_tooltip}
\alias{set_tooltip}
\title{Set tooltip}
\usage{
set_tooltip(widget, column)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the formatter is applied to.}
}
\description{
Set tooltip
}
\examples{
tabulator(airquality) |>
  set_formatter_traffic_light("Ozone") |>
  set_tooltip("Ozone")
}
