% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_top_calc}
\alias{set_top_calc}
\title{Add a calculation to a column}
\usage{
set_top_calc(
  widget,
  column,
  func = c("avg", "max", "min", "sum", "count", "unique"),
  precision = 2
)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the formatter is applied to.}

\item{func}{(character): The calculation function to be applied.}
}
\description{
Add a calculation to a column
}
