% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_options.R
\name{table_options}
\alias{table_options}
\title{Table Options}
\usage{
table_options(
  index = "id",
  columns = NULL,
  height = NULL,
  group_by = NULL,
  history = FALSE,
  layout = c("fitColumns", "fitData", "fitDataFill", "fitDataStretch", "fitDataTable"),
  selectable_rows = "highlight",
  header_visible = TRUE,
  movable_rows = FALSE,
  frozen_rows = NULL,
  row_height = NULL,
  add_row_pos = c("bottom", "top"),
  spreadsheet = FALSE,
  spreadsheet_rows = NULL,
  spreadsheet_columns = NULL,
  spreadsheet_column_definition = list(editor = "input"),
  ...
)
}
\description{
Table Options
}
