% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{Create a Tabulator Widget}
\usage{
tabulator(
  data,
  options = tabulator_options(),
  rtabulator_auto_columns = TRUE,
  editable = FALSE,
  theme = c("default", "midnight", "modern", "simple", "site", "bootstrap3",
    "bootstrap4", "bootstrap5", "bulma", "materialize", "semanticui"),
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{data}{(data.frame): description}

\item{options}{(list): description}

\item{rtabulator_auto_columns}{(bool): description}

\item{theme}{(character): description}

\item{width}{description}

\item{height}{description}

\item{elementId}{description}

\item{...}{Named arguments that are appended to options.}
}
\description{
Create a Tabulator Widget
}
