% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tabulator_data_as_df}
\alias{tabulator_data_as_df}
\title{Convert input data to data frame}
\usage{
tabulator_data_as_df(data)
}
\arguments{
\item{data}{(list): The input data returned in a Shiny app.}
}
\description{
Convert input data to data frame
}
