% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator_options.R
\name{tabulator_options}
\alias{tabulator_options}
\title{Tabulator Setup Options}
\usage{
tabulator_options(
  height = "311px",
  history = FALSE,
  columns = NULL,
  layout = c("fitColumns", "fitData", "fitDataFill", "fitDataStretch", "fitDataTable"),
  responsive_layout = FALSE,
  movable_columns = FALSE,
  header_visible = TRUE,
  row_height = NULL,
  add_row_pos = c("bottom", "top"),
  movable_rows = FALSE,
  resizable_rows = FALSE,
  frozen_rows = NULL,
  index = "id",
  group_by = NULL,
  edit_trigger_event = "dblclick",
  selectable_rows = "highlight",
  pagination = FALSE,
  pagination_size = 10,
  pagination_add_row = "page",
  spreadsheet = FALSE,
  spreadsheet_rows = NULL,
  spreadsheet_columns = NULL,
  spreadsheet_column_definition = NULL,
  spreadsheet_sheets = NULL,
  spreadsheet_sheet_tabs = NULL,
  ...
)
}
\arguments{
\item{height}{(character) The height of the table in px.}

\item{history}{(bool): Enable user interaction history functionality.}

\item{columns}{(list): Column definitions.
If set to \code{NULL} definitions are auto generated from table data.}

\item{layout}{(character): Layout mode of the table columns.}

\item{responsive_layout}{(bool): Automatically hide and show columns to fit
the width of the Tabulator element.}

\item{movable_columns}{(bool): Allow users to move and reorder rows.}

\item{header_visible}{(bool): Whether header is visible.}

\item{row_height}{(numeric): A fixed height of the rows.}

\item{add_row_pos}{(character): The table position where new rows are added.}

\item{movable_rows}{(bool): Allow users to move and reorder rows.}

\item{resizable_rows}{(bool): Allow user to resize rows.}

\item{frozen_rows}{(numeric): Number of frozen rows.}

\item{index}{(character): Field to be used as the unique index for each row.}

\item{group_by}{(character): Field to group rows by.}

\item{edit_trigger_event}{description}

\item{selectable_rows}{description}

\item{pagination}{(bool): description}

\item{pagination_size}{(integer): description}

\item{pagination_add_row}{(character): description}

\item{spreadsheet}{(bool): description}

\item{spreadsheet_rows}{(integer): description}

\item{spreadsheet_columns}{(integer): description}

\item{spreadsheet_column_definition}{description}

\item{spreadsheet_sheets}{(list): description}

\item{spreadsheet_sheet_tabs}{(bool): description}

\item{...}{Further options.}
}
\description{
Tabulator Setup Options
}
\seealso{
\url{https://tabulator.info/docs/6.2/options}
}
