% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator_context.R
\name{tabulatorContext}
\alias{tabulatorContext}
\title{Create a tabulator context object}
\usage{
tabulatorContext(output_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{output_id}{(character): A tabulator output id.}

\item{session}{A shiny session object.}
}
\value{
tabulator context object
}
\description{
Create a tabulator context object
}
\details{
A \code{tabulatorContext} object makes it possible to update your widget in a Shiny application.
}
\examples{
\dontrun{
tabulatorContext("table") |>
  trigger_download("csv")
}
}
