% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{titanic}
\alias{titanic}
\title{Titanic data set}
\usage{
titanic(col_select = NULL)
}
\arguments{
\item{col_select}{(character vector): Columns to select.}
}
\value{
data frame
}
\description{
Titanic data set
}
\examples{
\donttest{
titanic(c("PassengerId", "Age", "Pclass", "Fare")) |>
  head()
}

}
