% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{trigger_download}
\alias{trigger_download}
\title{Download table data}
\usage{
trigger_download(ctx, type = c("csv", "json", "xlsx"), file_name = NULL)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}

\item{type}{(character): csv, json or xlsx}

\item{file_name}{(character, \code{NULL}): File name.
If \code{NULL}, it is set to \code{"data.{type}"}.}
}
\value{
tabulator context object
}
\description{
Download table data
}
\details{
If you want to provide xlsx downloads, you need to include
the sheetjs html dependency: \code{tabulator(..., sheetjs = TRUE)}
}
\examples{
\dontrun{
tabulatorContext("table") |>
  trigger_download("csv", "table-data.csv")
}
}
