% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{trigger_get_data}
\alias{trigger_get_data}
\title{Submit data to R}
\usage{
trigger_get_data(ctx)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}
}
\value{
\link{tabulatorContext} object
}
\description{
Submit data to R
}
\examples{
\dontrun{
tabulatorContext("table") |>
  trigger_get_data()
}
}
