% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-calls.R
\name{trigger_get_sheet_data}
\alias{trigger_get_sheet_data}
\title{Submit sheet data to R}
\usage{
trigger_get_sheet_data(ctx)
}
\arguments{
\item{ctx}{A \code{\link[=tabulatorContext]{tabulatorContext()}} object.}
}
\value{
A \code{\link[=tabulatorContext]{tabulatorContext()}} object
}
\description{
Submit sheet data to R
}
\examples{
\donttest{
tabulatorContext("table") |>
  trigger_get_sheet_data()
}
}
