% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-calls.R
\name{undo}
\alias{undo}
\title{Undo changes}
\usage{
undo(ctx)
}
\arguments{
\item{ctx}{A \code{\link[=tabulatorContext]{tabulatorContext()}} object.}
}
\value{
A \code{\link[=tabulatorContext]{tabulatorContext()}} object
}
\description{
Undo changes
}
\examples{
\donttest{
tabulatorContext("table") |>
  undo()
}
}
