% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{add_row}
\alias{add_row}
\title{Add row to table}
\usage{
add_row(ctx, row = NULL)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}

\item{row}{(list | NULL): row data or \code{NULL} to add an empty row}
}
\value{
tabulator context object
}
\description{
Add row to table
}
