% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-calls.R
\name{add_row}
\alias{add_row}
\title{Add row to table}
\usage{
add_row(ctx, row = NULL)
}
\arguments{
\item{ctx}{A \code{\link[=tabulatorContext]{tabulatorContext()}} object.}

\item{row}{(list): row data or \code{NULL} to add an empty row}
}
\value{
A \code{\link[=tabulatorContext]{tabulatorContext()}} object
}
\description{
Add row to table
}
\examples{
\donttest{
tabulatorContext("table") |>
  add_row()
}
}
