% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{delete_selected_rows}
\alias{delete_selected_rows}
\title{Delete selected rows from the table}
\usage{
delete_selected_rows(ctx)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}
}
\description{
Delete selected rows from the table
}
