% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{delete_selected_rows}
\alias{delete_selected_rows}
\title{Delete selected rows from table}
\usage{
delete_selected_rows(ctx)
}
\arguments{
\item{ctx}{A \code{\link[=tabulatorContext]{tabulatorContext()}} object.}
}
\value{
A \code{\link[=tabulatorContext]{tabulatorContext()}} object
}
\description{
Delete selected rows from table
}
\examples{
\donttest{
tabulatorContext("table") |>
  delete_selected_rows()
}
}
