% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_data_frame}
\alias{list_to_data_frame}
\title{Convert a list of names lists to data frame}
\usage{
list_to_data_frame(x)
}
\arguments{
\item{x}{(list): A list of named lists.}
}
\description{
Convert a list of names lists to data frame
}
