% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_calculation}
\alias{set_calculation}
\title{Add a calculation to a column}
\usage{
set_calculation(
  widget,
  column,
  func = c("avg", "max", "min", "sum", "count", "unique"),
  precision = 2,
  pos = c("top", "bottom")
)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the \code{func} is applied to.}

\item{func}{(character): The calculation function to be applied
to the values of the \code{column}.}
}
\description{
Add a calculation to a column
}
\examples{
data_url <- "https://raw.githubusercontent.com/eodaGmbH/rtabulator/main/data-raw/titanic.csv"

tabulator(
  data_url,
  col_select = c("PassengerId", "Name", "Pclass", "Fare", "Survived"),
  pagination = T,
  column_calcs = "both"
) |>
  set_formatter_star("Pclass") |>
  set_formatter_tick_cross("Survived") |>
  set_formatter_progress("Fare", legend = TRUE, legend_align = "right", hoz_align = "right") |>
  set_calculation("Fare", func = "avg") |>
  set_calculation("Pclass", func = "unique", pos = "bottom")
}
