% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editors.R
\name{set_column_editor_input}
\alias{set_column_editor_input}
\title{Title}
\usage{
set_column_editor_input(
  widget,
  columns,
  search = TRUE,
  mask = "",
  select_contents = TRUE,
  elementAttributes = list(maxlength = 10)
)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{search}{Use search type input element with clear button}

\item{mask}{Apply a mask to the input to allow characters to be entered only in a certain order}

\item{select_contents}{When the editor is loaded select its text content}

\item{elementAttributes}{Set attributes directly on the input element}

\item{column}{The name of the column the formatter is applied to.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Title
}
\examples{
examples/editors/editor_input
}
