% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_formatter_datetime}
\alias{set_formatter_datetime}
\title{Set datetime formatter}
\usage{
set_formatter_datetime(
  widget,
  column,
  input_format = "yyyy-MM-dd hh:ss:mm",
  output_format = "yy/MM/dd",
  invalid_placeholder = "(invalid datetime)",
  timezone = NA
)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{column}{(character): The column the formatter is applied to.}

\item{input_format}{(character): The datetime input format.}

\item{output_format}{(character): The datetime output format.}

\item{invalid_placeholder}{(character): The value to be displayed if an invalid datetime is provided.}

\item{timezone}{(character): The timezone of the datetime.}
}
\description{
Set datetime formatter
}
\examples{
data <- data.frame(
  Person = c("Franz", "Ferdinand"),
  Birthday = c(
    "2024-12-06 22:00:10",
    "2023-06-07 14:12:45"
  )
)

tabulator(data) |>
  set_formatter_datetime("Birthday", output_format = "dd.MM.yyyy")
}
