% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_header_filter}
\alias{set_header_filter}
\title{Add header filter}
\usage{
set_header_filter(widget, columns = NULL)
}
\arguments{
\item{widget}{(\code{\link{tabulator}}) A tabulator widget.}

\item{columns}{(character vector): Columns the editor is applied to.
If set to \code{NULL}, the editor is applied to all columns.}
}
\description{
Add header filter
}
\examples{
data <- data.frame(
  age = c(10, 20, 40),
  first_name = c("Franz", "Ferdinand", "Julia"),
  last_name = c("Young", "Bowie", "Blackey")
)

tabulator(data) |>
  set_header_filter()
}
