% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{set_multi_column_header}
\alias{set_multi_column_header}
\title{Set multi column header}
\usage{
set_multi_column_header(widget, multi_columns)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{multi_columns}{(list): Multi column definitions.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Set multi column header
}
\examples{
multi_columns <- list(
  Sepal = c("Sepal_Length", "Sepal_Width"),
  Petal = c("Petal_Length", "Petal_Width")
)

tabulator(iris) |>
  set_multi_column_header(multi_columns)
}
