% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{set_tooltip}
\alias{set_tooltip}
\title{Set tooltip}
\usage{
set_tooltip(widget, columns)
}
\arguments{
\item{widget}{A \code{\link[=tabulator]{tabulator()}} HTML widget.}

\item{columns}{The names of the columns the formatter is applied to.}
}
\value{
The updated \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Set tooltip
}
\examples{
tabulator(airquality) |>
  set_formatter_traffic_light("Ozone") |>
  set_tooltip("Ozone")
}
