% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{spreadsheet_def}
\alias{spreadsheet_def}
\title{Create a spreadsheet definition}
\usage{
spreadsheet_def(title, key = NULL, data = NULL)
}
\arguments{
\item{title}{(character): The name of the spreadsheet.}

\item{key}{(character): Optional unique key of the spreadsheet.}

\item{data}{(list) The initial data of the spreadsheet.
Set to \code{NULL} to create an empty spreadsheet.}
}
\description{
Create a spreadsheet definition
}
