% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{spreadsheet_def}
\alias{spreadsheet_def}
\title{Create a spreadsheet definition}
\usage{
spreadsheet_def(title, key = NULL, data = NULL)
}
\arguments{
\item{title}{(character): The name of the spreadsheet.}

\item{key}{(character): Optional unique key of the spreadsheet.}

\item{data}{(list): The initial data of the spreadsheet.
Set to \code{NULL} to create an empty spreadsheet.}
}
\value{
list
}
\description{
Create a spreadsheet definition
}
\examples{
setup <- tabulator_options(
  spreadsheet = TRUE,
  spreadsheet_sheets = list(
    spreadsheet_def(
      title = "First",
      data = list(c(1, 2, 3))
    ),
    spreadsheet_def(
      title = "Second",
      data = list(c(4, 5, 6))
    )
  ),
  spreadsheet_sheet_tabs = TRUE
)

tabulator(data = NULL, setup, theme = "midnight")
}
