% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{Create a tabulator widget}
\usage{
tabulator(
  data,
  options = tabulator_options(),
  editable = FALSE,
  luxon = FALSE,
  sheetjs = FALSE,
  theme = c("default", "midnight", "modern", "simple", "site", "bootstrap3",
    "bootstrap4", "bootstrap5", "bulma", "materialize", "semanticui"),
  col_select = NULL,
  width = NULL,
  height = NULL,
  element_id = NULL,
  ...
)
}
\arguments{
\item{data}{(data.frame, character or list): In spreadsheet mode data needs to be a list or \code{NULL}
for an empty spreadsheet.}

\item{options}{(list): Setup options. See \code{\link{tabulator_options}}.}

\item{editable}{(bool): Whether the table is editable.}

\item{luxon}{(bool): Whether to add \href{https://moment.github.io/luxon/}{luxon} dependency,
which is required for \code{\link{set_formatter_datetime}}.}

\item{sheetjs}{(bool): Whether to add \href{https://sheetjs.com/}{sheetjs} dependency,
which is required for \code{\link{trigger_download}} to support Excel type downloads.}

\item{theme}{(character): Theme to apply to the table.}

\item{col_select}{(character vector) Columns to select.
Passed to \code{\link[readr]{read_csv}} if \code{data} is a file name.}

\item{width}{Width of the widget.}

\item{height}{Height of the widget.}

\item{element_id}{The unique ID of the widget.}

\item{...}{Named arguments that are appended to the \code{options} parameter.}
}
\value{
tabulator htmlwidget
}
\description{
Create a tabulator widget
}
\details{
Dots in column names are replaced by underscores.
}
\examples{
setup <- tabulator_options(
  spreadsheet = TRUE,
  spreadsheet_sheets = list(
    list(
      title = "First",
      data = list(c(1, 2, 3))
    ),
    list(
      title = "Second",
      data = list(c(4, 5, 6))
    )
  ),
  spreadsheet_sheet_tabs = TRUE
)

tabulator(data = NULL, setup, theme = "midnight")
}
