% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{<Add Title>}
\usage{
tabulator(
  data,
  options = tabulator_options(),
  rtabulator_auto_columns = TRUE,
  editable = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\description{
<Add Description>
}
