% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{Create a Tabulator Widget}
\usage{
tabulator(
  data,
  options = tabulator_options(),
  rtabulator_auto_columns = TRUE,
  editable = FALSE,
  theme = c("default", "midnight", "modern", "simple", "site", "bootstrap3",
    "bootstrap4", "bootstrap5", "bulma", "materialize", "semanticui"),
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{data}{(data.frame or list): In spreadsheet mode data needs to be a list or \code{NULL}
for an empty spreadsheet.}

\item{options}{(list): Setup options. See \code{\link{tabulator_options}}.}

\item{editable}{(bool): Whether the table is editable.}

\item{theme}{(character): Theme to apply to the table.}

\item{width}{Width of the widget.}

\item{height}{Height of the widget.}

\item{elementId}{description}

\item{...}{Named arguments that are appended to the \code{options} parameter.}
}
\description{
Create a Tabulator Widget
}
