% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator_options.R
\name{tabulator_options}
\alias{tabulator_options}
\title{Table Options
https://tabulator.info/docs/6.2/options}
\usage{
tabulator_options(
  height = NULL,
  history = FALSE,
  columns = NULL,
  layout = c("fitColumns", "fitData", "fitDataFill", "fitDataStretch", "fitDataTable"),
  responsive_layout = FALSE,
  movable_columns = FALSE,
  header_visible = TRUE,
  row_height = NULL,
  add_row_pos = c("bottom", "top"),
  movable_rows = FALSE,
  resizable_rows = FALSE,
  frozen_rows = NULL,
  index = "id",
  group_by = NULL,
  edit_trigger_event = "dblclick",
  selectable_rows = "highlight",
  pagination = FALSE,
  paginations_size = 10,
  pagination_add_row = "page",
  spreadsheet = FALSE,
  spreadsheet_rows = NULL,
  spreadsheet_columns = NULL,
  spreadsheet_column_definition = list(editor = "input"),
  spreadsheet_sheets = NULL,
  spreadsheet_sheet_tabs = NULL,
  ...
)
}
\description{
Table Options
https://tabulator.info/docs/6.2/options
}
