% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator_context.R
\name{tabulatorContext}
\alias{tabulatorContext}
\title{Create a Tabulator Context Object
A \code{tabulatorContext} object makes it possible to update your widget in a Shiny app.}
\usage{
tabulatorContext(output_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{output_id}{(character): A tabulator output id.}

\item{session}{shiny session object}
}
\description{
Create a Tabulator Context Object
A \code{tabulatorContext} object makes it possible to update your widget in a Shiny app.
}
