% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator-context.R
\name{tabulatorContext}
\alias{tabulatorContext}
\title{Context object to update a \code{tabulator()} HTML widget}
\usage{
tabulatorContext(output_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{output_id}{A tabulator output id set with \code{\link[=tabulatorOutput]{tabulatorOutput()}}.}

\item{session}{A shiny session object.}
}
\value{
A context object that can be used to interact with a \code{\link[=tabulator]{tabulator()}} HTML widget
}
\description{
Context object to update a \code{tabulator()} HTML widget
}
\details{
Makes it possible to update your \code{\link[=tabulator]{tabulator()}} HTML widget in Shiny apps.
}
\examples{
\donttest{
tabulatorContext("table") |>
  trigger_download("csv")
}
}
