% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{trigger_download}
\alias{trigger_download}
\title{Download table data}
\usage{
trigger_download(ctx, type = c("csv", "json"), file_name)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context}

\item{type}{(character): csv or json}

\item{file_name}{(character): file name}
}
\description{
Download table data
}
