% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{trigger_download}
\alias{trigger_download}
\title{Download table data}
\usage{
trigger_download(ctx, type = c("csv", "json", "xlsx"), file_name = NULL)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}

\item{type}{(character): csv, json or xlsx (needs sheetjs: \code{tabulator(..., sheetjs = TRUE)})}

\item{file_name}{(character, \code{NULL}): File name.}
}
\value{
tabulator context object
 If \code{NULL}, it is set to \code{"data.{type}"}.
}
\description{
Download table data
}
