% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_calls.R
\name{undo}
\alias{undo}
\title{Undo changes}
\usage{
undo(ctx)
}
\arguments{
\item{ctx}{(\code{\link{tabulatorContext}}): tabulator context object}
}
\value{
tabulator context object
}
\description{
Undo changes
}
\examples{
\dontrun{
tabulatorContext("table") |>
  undo()
}
}
